requirements.txt
# Core ML / NLP
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
transformers>=4.40.0
torch>=2.2.0

# LangChain
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
sse-starlette>=2.1.0

# Validation & config
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Document loading
pypdf>=4.2.0

# Caching
redis>=5.0.0

# Observability
structlog>=24.1.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Dev / lint
ruff>=0.4.0
mypy>=1.10.0
